Выводит шлейаут и шаблон Главной страницы
