+++ views/layouts/main.php
@@ -1,81 +1,61 @@
-<?php
-
-/* @var $this \yii\web\View */
-/* @var $content string */
-
-use app\assets\AppAsset;
-use app\widgets\Alert;
-use yii\bootstrap4\Breadcrumbs;
-use yii\bootstrap4\Html;
-use yii\bootstrap4\Nav;
-use yii\bootstrap4\NavBar;
-
-AppAsset::register($this);
-?>
-<?php $this->beginPage() ?>
 <!DOCTYPE html>
-<html lang="<?= Yii::$app->language ?>" class="h-100">
+<html lang="ru">
 <head>
-    <meta charset="<?= Yii::$app->charset ?>">
-    <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
-    <?php $this->registerCsrfMetaTags() ?>
-    <title><?= Html::encode($this->title) ?></title>
-    <?php $this->head() ?>
+    <meta charset="utf-8">
+    <meta name="viewport" content="width=device-width,initial-scale=1">
+    <title>Taskforce</title>
+    <link rel="stylesheet" href="css/style.css">
 </head>
-<body class="d-flex flex-column h-100">
-<?php $this->beginBody() ?>
+<body>
+<header class="page-header">
+    <nav class="main-nav">
+        <a href='#' class="header-logo">
+            <img class="logo-image" src="img/logotype.png" width=227 height=60 alt="taskforce">
+        </a>
+        <div class="nav-wrapper">
+            <ul class="nav-list">
+                <li class="list-item list-item--active">
+                    <a class="link link--nav" >Новое</a>
+                </li>
+                <li class="list-item">
+                    <a href="#" class="link link--nav" >Мои задания</a>
+                </li>
+                <li class="list-item">
+                    <a href="#" class="link link--nav" >Создать задание</a>
+                </li>
+                <li class="list-item">
+                    <a href="#" class="link link--nav" >Настройки</a>
+                </li>
+            </ul>
+        </div>
+    </nav>
+    <div class="user-block">
+        <a href="#">
+            <img class="user-photo" src="img/man-glasses.png" width="55" height="55" alt="Аватар">
+        </a>
+        <div class="user-menu">
+            <p class="user-name">Василий</p>
+            <div class="popup-head">
+                <ul class="popup-menu">
+                    <li class="menu-item">
+                        <a href="#" class="link">Настройки</a>
+                    </li>
+                    <li class="menu-item">
+                        <a href="#" class="link">Связаться с нами</a>
+                    </li>
+                    <li class="menu-item">
+                        <a href="#" class="link">Выход из системы</a>
+                    </li>
 
-<header>
-    <?php
-    NavBar::begin([
-        'brandLabel' => Yii::$app->name,
-        'brandUrl' => Yii::$app->homeUrl,
-        'options' => [
-            'class' => 'navbar navbar-expand-md navbar-dark bg-dark fixed-top',
-        ],
-    ]);
-    echo Nav::widget([
-        'options' => ['class' => 'navbar-nav'],
-        'items' => [
-            ['label' => 'Home', 'url' => ['/site/index']],
-            ['label' => 'About', 'url' => ['/site/about']],
-            ['label' => 'Contact', 'url' => ['/site/contact']],
-            Yii::$app->user->isGuest ? (
-                ['label' => 'Login', 'url' => ['/site/login']]
-            ) : (
-                '<li>'
-                . Html::beginForm(['/site/logout'], 'post', ['class' => 'form-inline'])
-                . Html::submitButton(
-                    'Logout (' . Yii::$app->user->identity->username . ')',
-                    ['class' => 'btn btn-link logout']
-                )
-                . Html::endForm()
-                . '</li>'
-            )
-        ],
-    ]);
-    NavBar::end();
-    ?>
+                </ul>
+            </div>
+        </div>
+    </div>
 </header>
-
-<main role="main" class="flex-shrink-0">
-    <div class="container">
-        <?= Breadcrumbs::widget([
-            'links' => isset($this->params['breadcrumbs']) ? $this->params['breadcrumbs'] : [],
-        ]) ?>
-        <?= Alert::widget() ?>
+<main class="main-content container">
+    <div class="main-container">
         <?= $content ?>
     </div>
 </main>
-
-<footer class="footer mt-auto py-3 text-muted">
-    <div class="container">
-        <p class="float-left">&copy; My Company <?= date('Y') ?></p>
-        <p class="float-right"><?= Yii::powered() ?></p>
-    </div>
-</footer>
-
-<?php $this->endBody() ?>
 </body>
 </html>
-<?php $this->endPage() ?>
